Move `Deprecated` to `version_utils`
